pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devmind"
version = "0.1.0"
description = "DevMind API"
authors = [
    {name = "Ray Sun", email = "xiaoquqi@gmail.com"}
]
requires-python = ">=3.8"
dependencies = [
    "django==5.1.4",
    "djangorestframework==3.15.2",
    "djangorestframework_simplejwt==5.3.1",
    "djangorestframework-camel-case==1.4.2",
    "dj-rest-auth==7.0.0",
    "drf-spectacular",
    "python-dotenv",
    "dj_database_url==2.3.0",
    "mysqlclient==2.1.1",  # Kept for backward compatibility (MySQL/MariaDB support)
    "psycopg2-binary>=2.9.9",  # PostgreSQL adapter (recommended)
    "celery==5.3.1",
    "django-celery-beat==2.7.0",
    "django_celery_results",
    "flower==2.0.1",
    "redis",
    "gunicorn==20.1.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "django-json-widget>=1.1.1",
    # Cloud billing providers
    "boto3>=1.34.0",  # AWS SDK
    "azure-identity>=1.15.0",  # Azure authentication
    "azure-mgmt-consumption>=10.0.0",  # Azure consumption management
    "azure-mgmt-resource>=23.0.0",  # Azure resource management
    "alibabacloud-bssopenapi20171214>=3.0.0",  # Alibaba Cloud BSS API
    "alibabacloud-tea-openapi>=0.3.0",  # Alibaba Cloud Tea OpenAPI
    "huaweicloudsdkcore>=3.1.0",  # Huawei Cloud SDK core
    "huaweicloudsdkbssintl>=3.1.0",  # Huawei Cloud BSS international
    "huaweicloudsdkbss>=3.1.0",  # Huawei Cloud BSS
    "django-allauth>=0.57.0",
    "django-cors-headers>=4.3.0",
    "django-extensions>=3.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.2",
    "black>=22.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]

[tool.setuptools]
packages = ["devmind"]

[project.scripts]
devmind-manager = "devmind.manage:main"